# Conflicts:
#	src/features/analysis/ui/analysisAnalysis/analysisAnalysis.tsx
#	src/features/analysis/ui/analysisPackageModal/analysisPackageModal.tsx
#	src/features/workTable/ui/workTable.tsx
#	src/pages/hospitalRegPage/ui/hospitalRegPage.tsx

+++ src/features/pakets/ui/userAnalysis.tsx
@@ -2,15 +2,16 @@ import React, {memo, useState} from 'react';
 import {useAppDispatch} from "../../../shared/lib/hooks/useAppDispatch/useAppDispatch";
 import {packetsActions, PacketsList, PacketsUserList} from "../../../entities/pakets";
 import {ConfirmModal} from "../../../shared/ui/confirm";
-import {IAnalysisProps} from "../../../entities/pakets/model/paketsSchema";
+import {IAnalysisProps, IUserPackets} from "../../../entities/pakets/model/paketsSchema";
 
 interface IUserAnalysis {
     item: IAnalysisProps[],
-    onDeleteAnalysisId: (arg: number) => void,
-    onDeleteAllAnalysis: () => void
+    total: string,
+    onDeleteAnalysisId?: (arg: number) => void,
+    onDeleteAllAnalysis?: () => void
 }
 
-export const UserAnalysis = memo(({item, onDeleteAnalysisId, onDeleteAllAnalysis}: IUserAnalysis) => {
+export const UserAnalysis = memo(({item, onDeleteAnalysisId, onDeleteAllAnalysis, total}: IUserAnalysis) => {
 
     const [isDeleteAnalysis, setIsDeleteAnalysis] = useState(false)
     const [isDeletePacket, setIsDeletePacket] = useState(false)
@@ -28,14 +29,18 @@ export const UserAnalysis = memo(({item, onDeleteAnalysisId, onDeleteAllAnalysis
         //     analysisId: isActiveAnalysis,
         //     packagePrice: price
         // }))
-        onDeleteAnalysisId(isActiveAnalysis)
-        setIsDeleteAnalysis(false)
+        if (onDeleteAnalysisId){
+            onDeleteAnalysisId(isActiveAnalysis)
+            setIsDeleteAnalysis(false)
+        }
     }
 
     const onDeletePacket = () => {
         // dispatch(deletePacket(isActivePacket))
-        onDeleteAllAnalysis()
-        setIsDeletePacket(false)
+        if (onDeleteAllAnalysis){
+            onDeleteAllAnalysis()
+            setIsDeletePacket(false)
+        }
     }
 
     const onClickAnalysis = (id: number) => {
@@ -53,6 +58,7 @@ export const UserAnalysis = memo(({item, onDeleteAnalysisId, onDeleteAllAnalysis
             <PacketsUserList
                 packet_id={1}
                 list={item}
+                total={total}
                 onDeleteAnalysis={onClickAnalysis}
                 onDeletePacket={onClickPacket}
             />
